stuff works, moving onto analytics

+++ conf/evolutions/default/2.sql
@@ -3,7 +3,7 @@
 # --- !Ups
 
 CREATE TABLE account (
-    id         integer NOT NULL PRIMARY KEY,
+    id         bigserial NOT NULL PRIMARY KEY,
     email      varchar NOT NULL UNIQUE,
     password   varchar NOT NULL,
     name       varchar NOT NULL,
